flux_infer::fixpoint_encoding

Struct FixpointCtxt

pub struct FixpointCtxt<'genv, 'tcx, T: Eq + Hash> {
    comments: Vec<String>,
    genv: GlobalEnv<'genv, 'tcx>,
    kvars: KVarGen,
    scx: SortEncodingCtxt,
    kcx: KVarEncodingCtxt,
    ecx: ExprEncodingCtxt<'genv, 'tcx>,
    tags: IndexVec<TagIdx, T>,
    tags_inv: UnordMap<T, TagIdx>,
}

Fields
comments: Vec<String>genv: GlobalEnv<'genv, 'tcx>kvars: KVarGenscx: SortEncodingCtxtkcx: KVarEncodingCtxtecx: ExprEncodingCtxt<'genv, 'tcx>tags: IndexVec<TagIdx, T>tags_inv: UnordMap<T, TagIdx>
Implementations

impl<'genv, 'tcx, Tag> FixpointCtxt<'genv, 'tcx, Tag>
where Tag: Hash + Eq + Copy,

pub(crate) fn fixpoint_to_solution( &mut self, sol: &FixpointSolution, ) -> Binder<Expr>

fn fixpoint_to_sort_ctor( &self, ctor: &SortCtor, ) -> Result<SortCtor, FixpointParseError>

pub(crate) fn fixpoint_to_sort( &self, fsort: &Sort, ) -> Result<Sort, FixpointParseError>

pub(crate) fn fixpoint_to_expr( &mut self, fexpr: &Expr, ) -> Result<Expr, FixpointParseError>

fn fixpoint_app_to_expr( &mut self, fhead: &Expr, fargs: &[Expr], ) -> Result<Expr, FixpointParseError>

impl<'genv, 'tcx, Tag> FixpointCtxt<'genv, 'tcx, Tag>
where Tag: Hash + Eq + Copy,

pub fn new( genv: GlobalEnv<'genv, 'tcx>, def_id: MaybeExternId, kvars: KVarGen, backend: Backend, ) -> Self

pub(crate) fn create_task( &mut self, def_id: MaybeExternId, constraint: Constraint, scrape_quals: bool, solver: SmtSolver, ) -> QueryResult<Task>

pub(crate) fn run_task( &mut self, cache: &mut FixQueryCache, def_id: MaybeExternId, kind: FixpointQueryKind, task: &Task, ) -> QueryResult<ParsedResult>

pub(crate) fn result_to_answer(&mut self, result: ParsedResult) -> Answer<Tag>

fn parse_kvar_solutions( &mut self, kvar_binds: &[KVarBind], ) -> FxIndexMap<KVid, FixpointSolution>

fn parse_kvar_solution(&mut self, expr: &str) -> FixpointSolution

fn is_assumed_constant(&self, const_decl: &ConstDecl) -> bool

fn extract_assumed_consts_aux( &self, cstr: Constraint, acc: &mut HashMap<GlobalVar, Expr>, ) -> Constraint

fn extract_assumed_consts( &self, cstr: Constraint, ) -> (HashMap<GlobalVar, Expr>, Constraint)

fn compute_const_deps( &self, constants: Vec<ConstDecl>, cstr: Constraint, ) -> (ConstDeps, Constraint)

pub fn generate_lean_files( self, def_id: MaybeExternId, task: Task, kvar_solutions: KVarSolutions, ) -> QueryResult

fn run_task_with_cache( genv: GlobalEnv<'_, '_>, task: &Task, def_id: DefId, kind: FixpointQueryKind, cache: &mut FixQueryCache, ) -> VerificationResult<TagIdx>

fn tag_idx(&mut self, tag: Tag) -> TagIdx
where Tag: Debug,

pub(crate) fn with_early_param(&mut self, param: &EarlyReftParam)

pub(crate) fn with_name_map<R>( &mut self, name: Name, provenance: NameProvenance, f: impl FnOnce(&mut Self, LocalVar) -> R, ) -> R

pub(crate) fn sort_to_fixpoint(&mut self, sort: &Sort) -> Sort

pub(crate) fn var_to_fixpoint(&self, var: &Var) -> Var

pub(crate) fn head_to_fixpoint( &mut self, expr: &Expr, mk_tag: impl Fn(Option<ESpan>) -> Tag + Copy, ) -> QueryResult<Constraint>
where Tag: Debug,

Encodes an expression in head position as a fixpoint::Constraint “peeling out” implications and foralls.

pub(crate) fn assumption_to_fixpoint( &mut self, pred: &Expr, ) -> QueryResult<(Vec<Bind>, Pred)>

Encodes an expression in assumptive position as a fixpoint::Pred. Returns the encoded predicate and a list of bindings produced by ANF-ing kvars.

fn assumption_to_fixpoint_aux( &mut self, expr: &Expr, bindings: &mut Vec<Bind>, preds: &mut Vec<Pred>, ) -> QueryResult

Auxiliary function to merge nested conjunctions in a single predicate

fn kvar_to_fixpoint( &mut self, kvar: &KVar, bindings: &mut Vec<Bind>, ) -> QueryResult<Pred>