flux_infer::infer
Struct InferCtxt

pub struct InferCtxt<'infcx, 'genv, 'tcx> {
    pub genv: GlobalEnv<'genv, 'tcx>,
    pub region_infcx: &'infcx InferCtxt<'tcx>,
    pub def_id: DefId,
    rcx: RefineCtxt<'infcx>,
    inner: &'infcx RefCell<InferCtxtInner>,
}

Fields
genv: GlobalEnv<'genv, 'tcx>region_infcx: &'infcx InferCtxt<'tcx>def_id: DefIdrcx: RefineCtxt<'infcx>inner: &'infcx RefCell<InferCtxtInner>
Implementations

impl<'infcx, 'genv, 'tcx> InferCtxt<'infcx, 'genv, 'tcx>

fn new( genv: GlobalEnv<'genv, 'tcx>, region_infcx: &'infcx InferCtxt<'tcx>, def_id: DefId, rcx: RefineCtxt<'infcx>, inner: &'infcx RefCell<InferCtxtInner>, ) -> Self

pub fn clean_subtree(&mut self, snapshot: &Snapshot)

pub fn change_item<'a>( &'a mut self, def_id: LocalDefId, region_infcx: &'a InferCtxt<'tcx>, ) -> InferCtxt<'a, 'genv, 'tcx>

pub fn change_root(&mut self, snapshot: &Snapshot) -> InferCtxt<'_, 'genv, 'tcx>

pub fn branch(&mut self) -> InferCtxt<'_, 'genv, 'tcx>

pub fn at(&mut self, span: Span) -> InferCtxtAt<'_, 'infcx, 'genv, 'tcx>

pub fn instantiate_refine_args( &mut self, callee_def_id: DefId, ) -> InferResult<List<Expr>>

pub fn instantiate_generic_args( &mut self, args: &[GenericArg], ) -> Vec<GenericArg>

pub fn fresh_infer_var(&self, sort: &Sort, mode: InferMode) -> Expr

pub fn fresh_infer_var_for_hole( &mut self, binders: &[BoundVariableKinds], kind: HoleKind, ) -> Expr

pub fn fresh_kvar( &self, binders: &[BoundVariableKinds], encoding: KVarEncoding, ) -> Expr

Generate a fresh kvar in the current scope. See KVarGen::fresh.

fn fresh_evar(&self) -> Expr

pub fn unify_exprs(&self, e1: &Expr, e2: &Expr)

fn enter_exists<T, U>( &mut self, exists: &Binder<T>, f: impl FnOnce(&mut Self, T) -> U, ) -> U
where T: TypeFoldable,

pub fn push_scope(&mut self)

The InferCtxt is a cursor into a tree. Some functions like define_var, assume_pred and by extension unpack advance the cursor. For example, defining a variable pushes a node as a child of the current node and then moves the cursor into that new node. Other functions, like subtyping or check_pred (typically the ones defined in InferCtxtAt) do not advance the cursor (from the caller’s perspective). For example, if you call subtyping a subtree is pushed under the current node, and the cursor is returned to where it was. That’s the purpose of infcx.branch: to “clone” the cursor such that the original one doesn’t get modified. infcx.pop_scope() and infcx.replace_evars(...) are supposed to be called when the infcx is on the same node where the push_scope was called, because replace_evars only replaces evars below that node. If pop_scope is called at a node further down (e.g. after calling unpack), the evars above that cursor will not be replaced. (see https://github.com/flux-rs/flux/pull/900#discussion_r1853052650)

pub fn pop_scope(&mut self) -> InferResult<EVarSol>

fn pop_scope_without_solving_evars(&mut self)

pub fn tcx(&self) -> TyCtxt<'tcx>