flux_infer::projections

Function normalize_projection_ty_with_rustc

fn normalize_projection_ty_with_rustc<'tcx>(
    genv: GlobalEnv<'_, 'tcx>,
    def_id: DefId,
    selcx: &mut SelectionContext<'_, 'tcx>,
    obligation: &AliasTy,
) -> QueryResult<(bool, SubsetTyCtor)>

Normalize an rty::AliasTy by converting it to rustc, normalizing it using rustc api, and then mapping the result back to rty. This will lose refinements and it should only be used to normalize sorts because they should only contain unrefined types. However, we are also using it as a hack to normalize types in cases where we fail to collect a candidate, this is unsound and should be removed.